src/requirement.txt
cn2an
python-dotenv
opencv-python
ocrmac
anthropic